# api/openapi.yaml
openapi: 3.0.3
info:
  title: LecodeKanban API
  version: 0.3.0
  description: |
    REST API of LecodeKanban. Source of truth for Go DTOs and the TypeScript client (`make gen`).

    **Auth:** httpOnly cookies — `lk_at` (access, ~15 min) and `lk_rt` (rotating refresh,
    path-scoped to `/api/v1/auth`). **CSRF:** every non-GET request must echo the `lk_csrf`
    cookie in the `X-CSRF-Token` header (obtain it via `GET /auth/csrf`).

    **Errors** are `{"error": {"code", "message", "fields"?, "meta"?}}`. `code` is stable and
    translated by clients; `message` is an English developer hint. See `GET /i18n/error-codes`.
servers:
  - url: /api/v1
...
paths:
...
  /projects/{projectId}/board:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [boards]
      operationId: getBoard
      responses:
        "200":
          description: Board with ordered columns
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Board" }
...
  /projects/{projectId}/board/columns:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    post:
      tags: [boards]
      operationId: createColumn
      description: Adds a custom column right after the last column of its status.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ColumnInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardColumn" }
        "422": { $ref: "#/components/responses/Error" }
  /columns/{columnId}:
    parameters: [{ $ref: "#/components/parameters/ColumnId" }]
    patch:
      tags: [boards]
      operationId: updateColumn
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ColumnPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardColumn" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [boards]
      operationId: deleteColumn
      description: Only empty columns can be deleted, and every status keeps at least one column.
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /columns/{columnId}/move:
    parameters: [{ $ref: "#/components/parameters/ColumnId" }]
    post:
      tags: [boards]
      operationId: moveColumn
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Neighbours" }
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardColumn" }
  /workspaces/{workspaceId}/views:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [boards]
      operationId: listSavedViews
      responses:
        "200":
          description: The caller's saved views
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SavedView" } }
    post:
      tags: [boards]
      operationId: createSavedView
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SavedViewInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedView" }
        "422": { $ref: "#/components/responses/Error" }
  /views/{viewId}:
    parameters: [{ $ref: "#/components/parameters/ViewId" }]
    patch:
      tags: [boards]
      operationId: updateSavedView
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SavedViewPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedView" }
    delete:
      tags: [boards]
      operationId: deleteSavedView
      responses:
        "204": { description: Deleted }
...
  /cards/{cardId}/checklist:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    get:
      tags: [cards]
      operationId: listChecklist
      responses:
        "200":
          description: Items in order
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ChecklistItem" } }
    post:
      tags: [cards]
      operationId: addChecklistItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChecklistItemInput" }
      responses:
        "201":
          description: Created (card progress is recomputed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChecklistItem" }
  /checklist-items/{itemId}:
    parameters: [{ $ref: "#/components/parameters/ItemId" }]
    patch:
      tags: [cards]
      operationId: updateChecklistItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChecklistItemPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChecklistItem" }
    delete:
      tags: [cards]
      operationId: deleteChecklistItem
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: lk_at }
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CardId:
      name: cardId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ColumnId: { name: columnId, in: path, required: true, schema: { type: string, format: uuid } }
    ViewId: { name: viewId, in: path, required: true, schema: { type: string, format: uuid } }
...
    ItemId: { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: auth.invalid_credentials }
            message: { type: string }
            fields:
              type: array
              items: { $ref: "#/components/schemas/FieldError" }
            meta:
              type: object
              additionalProperties: true
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, example: min_length }
        params: { type: object, additionalProperties: true }
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, in_review, done]
...
    BoardColumn:
      type: object
      required: [id, name, status, position, wipLimit]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        position: { type: string }
        wipLimit: { type: integer, nullable: true }
    Board:
      type: object
      required: [id, projectId, name, columns]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        name: { type: string }
        columns: { type: array, items: { $ref: "#/components/schemas/BoardColumn" } }
...
    ChecklistItem:
      type: object
      required: [id, text, done, position, completedAt]
      properties:
        id: { type: string, format: uuid }
        text: { type: string }
        done: { type: boolean }
        position: { type: string }
        completedAt: { type: string, format: date-time, nullable: true }
    ChecklistItemInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 300 }
    ChecklistItemPatch:
      type: object
      description: Any subset; `move` reorders between neighbours.
      properties:
        text: { type: string, minLength: 1, maxLength: 300 }
        done: { type: boolean }
        move: { $ref: "#/components/schemas/Neighbours" }
    Neighbours:
      type: object
      properties:
        afterId: { type: string, format: uuid, nullable: true, description: Item that will be directly before }
        beforeId: { type: string, format: uuid, nullable: true, description: Item that will be directly after }
    ColumnInput:
      type: object
      required: [name, status]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        status: { $ref: "#/components/schemas/TaskStatus" }
        wipLimit: { type: integer, minimum: 1, maximum: 999, nullable: true }
    ColumnPatch:
      type: object
      description: Absent fields are unchanged; wipLimit null removes the limit.
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        wipLimit: { type: integer, minimum: 1, maximum: 999, nullable: true }
    SavedView:
      type: object
      required: [id, name, config, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        config: { type: object, additionalProperties: true, description: "Opaque view configuration (filters, swimlanes, project)" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SavedViewInput:
      type: object
      required: [name, config]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        config: { type: object, additionalProperties: true }
    SavedViewPatch:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        config: { type: object, additionalProperties: true }
...
security:
  - cookieAuth: []